requirements.txt
flask
mysql-connector-python
numpy
pandas
